pyproject.toml
[project]
name = "medical-edu-explainer"
version = "0.1.0"
description = "Transform medical education PDFs into AI explainer video scripts"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "pymupdf>=1.24.0",
    "pyairtable>=2.3.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    "elevenlabs>=1.0.0",
    "google-genai>=1.0.0",
    "streamlit>=1.42.0",
    "pytest>=8.3.0",
]

[project.scripts]
medexplainer = "run:main"
